# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/management/places/{id}:
    put:
      tags:
      - place-management-rest-controller
      summary: Edit existing place
      operationId: editPlace
      parameters:
      - name: id
        in: path
        description: id of place to be edited
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceDTO'
        required: true
      responses:
        "200":
          description: Place was correctly edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "404":
          description: Place not found
        "403":
          description: Not authorized
    delete:
      tags:
      - place-management-rest-controller
      summary: Delete place
      operationId: deletePlace
      parameters:
      - name: id
        in: path
        description: id of place to be deleted
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Place was correctly deleted
        "404":
          description: Place not found
        "403":
          description: Not authorized
  /api/management/places/{id}/image:
    put:
      tags:
      - place-management-rest-controller
      summary: Edit existing place's image
      operationId: editImage_2
      parameters:
      - name: id
        in: path
        description: id of place to edit its image
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to be uploaded
                  format: binary
      responses:
        "200":
          description: Place's image was edited correctly
          content:
            image/jpeg:
              schema:
                $ref: '#/components/schemas/Place'
        "404":
          description: Place not found
        "403":
          description: Not authorized
...
  /api/itineraries/{id}/image:
    get:
      tags:
      - itinerary-rest-controller
      summary: Returns the image of the desired Itinerary.
      operationId: downloadImage_2
      parameters:
      - name: id
        in: path
        description: itinerary id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Itinerary's image.
        "200":
          description: Sucessfully returned the Itinerary's image.
          content:
            image/jpeg: {}
    put:
      tags:
      - itinerary-rest-controller
      summary: Edit the image of an itinerary
      operationId: editItineraryImage
      parameters:
      - name: id
        in: path
        description: id of the itinerary you want to add an image to
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/jpeg:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/png:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
      responses:
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "404":
          description: Itinerary not found
        "200":
          description: Image edited successfully
          content:
            image/jpeg: {}
...
  /api/management/places:
    get:
      tags:
      - place-management-rest-controller
      summary: Get a page with all places
      operationId: getPlaces
      parameters:
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Places not found
        "200":
          description: Places page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "403":
          description: Not authorized
    post:
      tags:
      - place-management-rest-controller
      summary: Create a new place
      operationId: createPlace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceDTO'
        required: true
      responses:
        "201":
          description: New place was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "403":
          description: Not authorized
...
  /api/itineraries/{itineraryId}/places:
    post:
      tags:
      - itinerary-rest-controller
      summary: Add a place to an owned itinerary
      operationId: editPlaces
      parameters:
      - name: itineraryId
        in: path
        description: id of the itinerary you want to add a place to
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              description: id of the place you want to add to the itinerary
              format: int64
        required: true
      responses:
        "200":
          description: Place added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "404":
          description: Itinerary or place not found
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Returns the profile image of the desired user
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        description: user id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the user image
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing User's image.
...
  /api/places:
    get:
      tags:
      - place-rest-controller
      summary: Searches for Places in the database. Filters can be applied optionally.
      operationId: places
      parameters:
      - name: name
        in: query
        description: search query
        required: false
        schema:
          type: string
          default: ""
      - name: sort
        in: query
        description: "sorting type: id, name, views"
        required: false
        schema:
          type: string
          default: id
      - name: order
        in: query
        description: order specifier
        required: false
        schema:
          type: string
          default: DESC
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: Successfully searched the desired Places.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "400":
          description: Invalid arguments.
  /api/places/{id}:
    get:
      tags:
      - place-rest-controller
      summary: Searches for Places in the database. Filters can be applied optionally.
      operationId: place
      parameters:
      - name: id
        in: path
        description: place id
        required: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: itineraries page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: Successfully searched the desired Place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetailsDTO'
        "400":
          description: Invalid arguments.
  /api/places/{id}/image:
    get:
      tags:
      - place-rest-controller
      summary: Returns the image of the desired Place.
      operationId: downloadImage_1
      parameters:
      - name: id
        in: path
        description: place id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Place's image.
        "200":
          description: Sucessfully returned the Place's image.
          content:
            image/jpeg: {}
  /api/itineraries/{id}/places:
    get:
      tags:
      - itinerary-rest-controller
      summary: Get Detailed information about a specific Itinerary.
      operationId: getPlacesInUserItinerary
      parameters:
      - name: id
        in: path
        description: id of the itinerary to view places of
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: places page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Requested a non-existing Itinerary.
        "200":
          description: Successfully searched the desired Places frmo an Itinerary.
        "403":
          description: Invalid permissions to request places from a private Itinerary.
        "400":
          description: Invalid arguments.
...
  /api/destinations/{id}/places:
    get:
      tags:
      - destination-rest-controller
      summary: Get Places from a specific Destination.
      operationId: destinationPlaces
      parameters:
      - name: id
        in: path
        description: destination id
        required: true
        schema:
          type: integer
          format: int32
      - name: placesPage
        in: query
        description: places page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Requested a non-existing Destination.
        "200":
          description: Successfully searched the desired destinations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationDetailsDTO'
        "400":
          description: Invalid arguments.
  /api/destinations/{id}/image:
    get:
      tags:
      - destination-rest-controller
      summary: Returns the image of the desired Destination.
      operationId: downloadImage_3
      parameters:
      - name: id
        in: path
        description: destination id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the Destination's image.
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing Destination's image.
  /api/itineraries/{itineraryId}/places/{placeId}:
    delete:
      tags:
      - itinerary-rest-controller
      summary: Remove a place from an owned itinerary
      operationId: deletePlacesFromItinerary
      parameters:
      - name: itineraryId
        in: path
        description: id of the itinerary to modify
        required: true
        schema:
          type: integer
          format: int64
      - name: placeId
        in: path
        description: id of the place to remove
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "404":
          description: Itinerary or place not found
        "200":
          description: Place removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        nationality:
          type: string
        image:
          type: boolean
        roles:
          type: array
          items:
            type: string
...
    PlaceDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        destination:
          type: string
      description: new page's information
    Destination:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        flagCode:
          type: string
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
    Place:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        destination:
          $ref: '#/components/schemas/Destination'
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
...
    Itinerary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        public:
          type: boolean
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
...
    PageItinerary:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Itinerary'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    PlaceDetailsDTO:
      type: object
      properties:
        place:
          $ref: '#/components/schemas/Place'
        itineraries:
          $ref: '#/components/schemas/PageItinerary'
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PagePlace:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
...
    DestinationDetailsDTO:
      type: object
      properties:
        destination:
          $ref: '#/components/schemas/Destination'
        places:
          $ref: '#/components/schemas/PagePlace'