# backend/src/main/java/com/tripscanner/TripScanner/model/Itinerary.java
package com.tripscanner.TripScanner.model;

import com.fasterxml.jackson.annotation.JsonIgnore;
import com.tripscanner.TripScanner.model.rest.ItineraryDTO;
import org.hibernate.engine.jdbc.BlobProxy;
import org.springframework.core.io.InputStreamResource;
import org.springframework.core.io.Resource;

import javax.persistence.*;
import java.io.IOException;
import java.io.InputStream;
import java.sql.Blob;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.List;

@Entity
public class Itinerary implements Information {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id = null;

    private String name;

    @Column(columnDefinition = "TEXT")
    private String description;

    private Long views;

    private boolean image;

    private boolean isPublic;

    @Lob
    @JsonIgnore
    private Blob imageFile;

    @ManyToMany
    @JsonIgnore
    private List<Place> places;

    @ManyToOne
    private User user;

    @OneToMany(mappedBy = "itinerary", cascade = CascadeType.ALL, orphanRemoval = true)
    @JsonIgnore
    private List<Review> reviews;

    public Itinerary() {
    }

    public Itinerary(String name, String description, User user, boolean isPublic) {
        super();
        this.name = name;
        this.description = description;
        this.user = user;
        this.isPublic = isPublic;
        this.setImage(false);
        this.setViews(0L);
    }

    public Itinerary(ItineraryDTO itineraryDTO, User user, boolean isPublic) {
        super();
        this.name = itineraryDTO.getName();
        this.description = itineraryDTO.getDescription();
        this.user = user;
        this.isPublic = isPublic;
        this.setImage(false);
        this.setViews(0L);
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    @Override
    public String getType() {
        return "Itinerary";
    }

    @Override
    public String getTypeLowercase() {
        return getType().toLowerCase();
    }

    @Override
    public String getFlag() {
        if (places == null || places.isEmpty()) {
            return "https://flagicons.lipis.dev/flags/4x3/xx.svg";
        } else {
            return places.get(0).getFlag();
        }
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Long getViews() {
        return views;
    }

    public void setViews(Long views) {
        this.views = views;
    }

    @Override
    public boolean isImage() {
        return image;
    }

    public void setImage(boolean image) {
        this.image = image;
    }

    @Override
    public Blob getImageFile() {
        return imageFile;
    }

    public void setImageFile(Blob imageFile) {
        this.imageFile = imageFile;
    }

    public List<Place> getPlaces() {
        return places;
    }

    public void setPlaces(List<Place> places) {
        this.places = places;
    }

    public User getUser() {
        return user;
    }

    public void setUser(User user) {
        this.user = user;
    }

    public List<Review> getReviews() {
        return reviews;
    }

    public void setReviews(List<Review> reviews) {
        this.reviews = reviews;
    }

    public Boolean isPublic() {
        return isPublic;
    }

    public void setPublic(boolean aPublic) {
        this.isPublic = aPublic;
    }

    public Itinerary copy(User newUser)  {
        Itinerary toReturn = new Itinerary(this.name, this.description, newUser, newUser.getRoles().contains("ADMIN"));
        try {
            byte[] imageBytes = this.getImageFile().getBytes(1L, (int) this.getImageFile().length());
            toReturn.setImageFile(BlobProxy.generateProxy(imageBytes));
            toReturn.setImage(true);
        } catch (SQLException e) {
            toReturn.setImage(false);
        }
        List<Place> placeCopy = new ArrayList<>();

        for (Place item : this.getPlaces()) {
            placeCopy.add(item);
        }

        toReturn.setPlaces(placeCopy);
        return toReturn;
    }

    public boolean hasName() {
        return !this.name.isBlank() || !this.name.isEmpty() || !(this.name == null);
    }

    public boolean hasDescription() {
        return !(this.description == null);
    }

}